pyproject.toml
[project]
name = "bangi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-httpauth>=4.8.0",
    "flask-smorest>=0.46.1",
    "gunicorn>=23.0.0",
    "peewee>=3.18.2",
    "peewee-migrate>=1.13.0",
    "pyhumps>=3.8.0",
    "pytest-dotenv>=0.5.2",
    "python-json-logger>=3.3.0",
    "wireup>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mysql>=3.1.0",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
